Refactor roles/reports and add option to upload to s3 (dry-run)
Instead of having a gigantic command in the crontab we now have .sh
scripts with all the commands and cron just calls it.

Added options to run aws-cli via docker and upload the reports generated
to S3/Spaces. For now it runs with the --dryrun flag so we can test it
before enabling for real.